Move KafkaRaftLog to internals package
